+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 - *resource stream* feature
 - `aws`, `sql` blob storage options; `quest-py` extras
 - fixed historian storage memory leak, storages clearing properly
+- added SIGINT handler
 
 ## 0.2.2